+++ DAL/DBMODEL/SaleOrder.cs
@@ -16,7 +16,7 @@ namespace DAL.DBMODEL
         public int RestaurantId { get; set; }
         public int UserId { get; set; }
         public int OrderCount { get; set; }
-        public long Total { get; set; }
+        public decimal Total { get; set; }
         public DateTime Date { get; set; }
         public DateTime? OrderDate { get; set; }
         public string Description { get; set; }
@@ -25,7 +25,7 @@ namespace DAL.DBMODEL
         public string CouponType { get; set; }
         public string CouponAppliesTo { get; set; }
         public string CouponCategories { get; set; }
-        public long? DiscountAmount { get; set; }
+        public decimal? DiscountAmount { get; set; }
         public string DiscountDesc { get; set; }
         public string PaymentMode { get; set; }
         public string PaymentStatus { get; set; }
@@ -36,10 +36,10 @@ namespace DAL.DBMODEL
         public string CustomerPhone { get; set; }
         public string IsUpdated { get; set; }
         public string IsDeleted { get; set; }
-        public long CashAmount { get; set; }
-        public long OnlineAmount { get; set; }
+        public decimal CashAmount { get; set; }
+        public decimal OnlineAmount { get; set; }
         public int IsPrinted { get; set; }
-        public long? ServiceCharge { get; set; }
+        public decimal? ServiceCharge { get; set; }
         public decimal? DeliveryCharges { get; set; }
 
         public virtual Customer Customer { get; set; }
