+++ DAL/DBMODEL/SaleOrderDetail.cs
@@ -24,8 +24,8 @@ namespace DAL.DBMODEL
         public int KitchenLines { get; set; }
         public string ItemComments { get; set; }
         public int ItemQty { get; set; }
-        public long? ItemPrice { get; set; }
-        public long? PurchasePrice { get; set; }
+        public decimal? ItemPrice { get; set; }
+        public decimal? PurchasePrice { get; set; }
         public int ItemIndex { get; set; }
         public int? BillNo { get; set; }
         public string IsUpdated { get; set; }
