+++ EZYPOS/UserControls/Report/UserControlSaleOrderReport.xaml.cs
@@ -59,7 +59,7 @@ namespace EZYPOS.UserControls.Report
                 var Items = DB.SaleOrder.GetAll().Where(x => x.OrderDate >= Sdate && x.OrderDate <= Edate).ToList();
                 string CustomerName = "";
                 string EmployeeName = "";
-                long GrandTotal = 0;
+                decimal GrandTotal = 0;
                 myList.Clear();
                 foreach(var item in Items)
                 {
@@ -100,7 +100,7 @@ namespace EZYPOS.UserControls.Report
                             var Items = DB.SaleOrder.GetAll().Where(x => x.OrderDate >= Sdate && x.OrderDate <= Edate).ToList();
                             string CustomerName = "";
                             string EmployeeName = "";
-                            long GrandTotal = 0;
+                            decimal GrandTotal = 0;
                             myList.Clear();
                             foreach (var item in Items)
                             {
@@ -124,7 +124,7 @@ namespace EZYPOS.UserControls.Report
                             var Items = DB.SaleOrder.GetAll().Where(x => x.OrderDate >= Sdate && x.OrderDate <= Edate).ToList();
                             string CustomerName = "";
                             string EmployeeName = "";
-                            long GrandTotal = 0;
+                            decimal GrandTotal = 0;
                             myList.Clear();
                             foreach (var item in Items)
                             {
@@ -156,7 +156,7 @@ namespace EZYPOS.UserControls.Report
                             string CustomerName = "";
                             string EmployeeName = "";
                             string PaymentMode = "";
-                            long GrandTotal = 0;
+                            decimal GrandTotal = 0;
                             myList.Clear();
                             foreach (var item in Items)
                             {
