+++ EZYPOS/UserControls/Transaction/UserControlSaleItem.xaml.cs
@@ -57,7 +57,7 @@ namespace EZYPOS.UserControls.Transaction
             //if (listBoxItemCart.Items.Count != 0)
             {
 
-                order.DeliverCharges = Convert.ToDouble(parameter);
+                order.DeliverCharges = Convert.ToDecimal(parameter);
                 UpdateBillSummary();
             }
 
@@ -768,7 +768,7 @@ namespace EZYPOS.UserControls.Transaction
                 Discount popup = new Discount();
                 if (popup.ShowDialog() == true)
                 {
-                    double digit = Convert.ToDouble(popup.pin);
+                    decimal digit = Convert.ToDecimal(popup.pin);
                     if (popup.DiscountType.SelectedIndex == 0)
                     {
                         ActiveSession.order_Discount_percentage = 0;
